requirements.txt
cryptography==42.0.5
PyYAML==6.0.1
waitress==3.0.0
fastapi~=0.112.1
Jinja2~=3.1.3
redis~=5.0.8
fastapi-limiter~=0.1.6
python-multipart~=0.0.9
uvicorn~=0.30.6